+++ lib/fog/azurerm/requests/compute/create_availability_set.rb
@@ -3,10 +3,18 @@ module Fog
     class AzureRM
       # This class provides the actual implementation for service calls.
       class Real
-        def create_availability_set(resource_group, name, location, is_managed = false)
+        def create_availability_set(availability_set_params)
+          name = availability_set_params[:name]
+          location = availability_set_params[:location]
+          resource_group = availability_set_params[:resource_group]
+          fault_domain_count = availability_set_params[:platform_fault_domain_count]
+          update_domain_count = availability_set_params[:platform_update_domain_count]
+          is_managed = availability_set_params[:is_managed].nil? ? false : availability_set_params[:is_managed]
+
           msg = "Creating Availability Set '#{name}' in #{location} region."
           Fog::Logger.debug msg
-          avail_set_params = get_availability_set_properties(location, is_managed)
+          avail_set_params = get_availability_set_properties(location, fault_domain_count, update_domain_count, is_managed)
+
           begin
             availability_set = @compute_mgmt_client.availability_sets.create_or_update(resource_group, name, avail_set_params)
           rescue MsRestAzure::AzureOperationError => e
@@ -17,14 +25,12 @@ module Fog
         end
 
         # create the properties object for creating availability sets
-        def get_availability_set_properties(location, is_managed)
+        def get_availability_set_properties(location, fault_domain_count, update_domain_count, is_managed)
           avail_set = Azure::ARM::Compute::Models::AvailabilitySet.new
-          avail_set.platform_fault_domain_count = FAULT_DOMAIN_COUNT
-          avail_set.platform_update_domain_count = UPDATE_DOMAIN_COUNT
-          avail_set.virtual_machines = []
-          avail_set.statuses = []
           avail_set.location = location
           avail_set.sku = create_availability_set_sku(is_managed)
+          avail_set.platform_fault_domain_count = fault_domain_count.nil? ? FAULT_DOMAIN_COUNT : fault_domain_count
+          avail_set.platform_update_domain_count = update_domain_count.nil? ? UPDATE_DOMAIN_COUNT : update_domain_count
           avail_set
         end
 
@@ -34,13 +40,14 @@ module Fog
           sku
         end
       end
+
       # This class provides the mock implementation for unit tests.
       class Mock
-        def create_availability_set(resource_group, name, location)
+        def create_availability_set(availability_set_params)
           {
-            'location' => location,
-            'id' => "/subscriptions/########-####-####-####-############/resourceGroups/#{resource_group}/providers/Microsoft.Compute/availabilitySets/#{name}",
-            'name' => name,
+            'location' => availability_set_params[:location],
+            'id' => "/subscriptions/########-####-####-####-############/resourceGroups/#{availability_set_params[:resource_group]}/providers/Microsoft.Compute/availabilitySets/#{availability_set_params[:name]}",
+            'name' => availability_set_params[:name],
             'type' => 'Microsoft.Compute/availabilitySets',
             'properties' =>
             {