+++ test/api_stub/requests/compute/availability_set.rb
@@ -11,7 +11,7 @@ module ApiStub
                     "location":"westus",
                     "tags": {},
                     "platformUpdateDomainCount": 5,
-                    "platformFaultDomainCount":3,
+                    "platformFaultDomainCount": 3,
                     "virtualMachines":[],
                     "sku":{
                       "name":"Classic"
@@ -29,11 +29,41 @@ module ApiStub
                     "location":"westus",
                     "tags": {},
                     "platformUpdateDomainCount": 5,
-                    "platformFaultDomainCount":3,
+                    "platformFaultDomainCount": 2,
                     "virtualMachines":[],
                     "sku":{
                       "name":"Aligned"
                     }
+                  }'
+          availability_set_mapper = Azure::ARM::Compute::Models::AvailabilitySet.mapper
+          compute_client.deserialize(availability_set_mapper, Fog::JSON.decode(body), 'result.body')
+        end
+
+        def self.create_availability_set_response(compute_client)
+          body = '{
+                    "id":"/subscriptions/{subscription-id}/resourceGroups/myrg1/providers/Microsoft.Compute/availabilitySets/avset1",
+                    "name":"myavset1",
+                    "type":"Microsoft.Compute/availabilitySets",
+                    "location":"westus",
+                    "tags": {},
+                    "platformUpdateDomainCount": 5,
+                    "platformFaultDomainCount": 3,
+                    "virtualMachines":[]
+                  }'
+          availability_set_mapper = Azure::ARM::Compute::Models::AvailabilitySet.mapper
+          compute_client.deserialize(availability_set_mapper, Fog::JSON.decode(body), 'result.body')
+        end
+
+        def self.create_custom_availability_set_response(compute_client)
+          body = '{
+                    "id":"/subscriptions/{subscription-id}/resourceGroups/myrg1/providers/Microsoft.Compute/availabilitySets/avset1",
+                    "name":"myavset1",
+                    "type":"Microsoft.Compute/availabilitySets",
+                    "location":"westus",
+                    "tags": {},
+                    "platformUpdateDomainCount": 10,
+                    "platformFaultDomainCount": 3,
+                    "virtualMachines":[]
                   }'
           availability_set_mapper = Azure::ARM::Compute::Models::AvailabilitySet.mapper
           compute_client.deserialize(availability_set_mapper, Fog::JSON.decode(body), 'result.body')