+++ test/requests/compute/test_create_availability_set.rb
@@ -8,38 +8,104 @@ class TestCreateAvailabilitySet < Minitest::Test
     @availability_sets = @client.availability_sets
   end
 
-  def test_create_unmanaged_availability_set_success
-    mocked_response = ApiStub::Requests::Compute::AvailabilitySet.create_unmanaged_availability_set_response(@client)
+  def test_create_availability_set_success
+    mocked_response = ApiStub::Requests::Compute::AvailabilitySet.create_availability_set_response(@client)
+    avail_set_params =
+      {
+        name: 'myavset1',
+        location: 'west us',
+        resource_group: 'fog-test-rg'
+      }
+
     @availability_sets.stub :validate_params, true do
       @availability_sets.stub :create_or_update, mocked_response do
-        assert_equal @service.create_availability_set('fog-test-rg', 'myavset1', 'west us', false), mocked_response
+        assert_equal @service.create_availability_set(avail_set_params), mocked_response
+      end
+    end
+  end
+
+  def test_create_custom_availability_set_success
+    mocked_response = ApiStub::Requests::Compute::AvailabilitySet.create_custom_availability_set_response(@client)
+    avail_set_params =
+      {
+        name: 'myavset1',
+        location: 'west us',
+        resource_group: 'fog-test-rg',
+        platform_fault_domain_count: 3,
+        platform_update_domain_count: 10
+      }
+
+    @availability_sets.stub :validate_params, true do
+      @availability_sets.stub :create_or_update, mocked_response do
+        assert_equal @service.create_availability_set(avail_set_params), mocked_response
+      end
+    end
+  end
+
+  def test_create_unmanaged_availability_set_success
+    mocked_response = ApiStub::Requests::Compute::AvailabilitySet.create_unmanaged_availability_set_response(@client)
+    avail_set_params =
+      {
+        name: 'myavset1',
+        location: 'west us',
+        resource_group: 'fog-test-rg',
+        is_managed: false
+      }
+
+    @availability_sets.stub :validate_params, true do
+      @availability_sets.stub :create_or_update, mocked_response do
+        assert_equal @service.create_availability_set(avail_set_params), mocked_response
       end
     end
   end
 
   def test_create_unmanaged_availability_set_failure
     response = proc { raise MsRestAzure::AzureOperationError.new(nil, nil, 'error' => { 'message' => 'mocked exception' }) }
+    avail_set_params =
+      {
+        name: 'myavset1',
+        location: 'west us',
+        resource_group: 'fog-test-rg',
+        is_managed: false
+      }
+
     @availability_sets.stub :validate_params, true do
       @availability_sets.stub :create_or_update, response do
-        assert_raises(RuntimeError) { @service.create_availability_set('fog-test-rg', 'fog-test-as', 'west us', false) }
+        assert_raises(RuntimeError) { @service.create_availability_set(avail_set_params) }
       end
     end
   end
 
   def test_create_managed_availability_set_success
     mocked_response = ApiStub::Requests::Compute::AvailabilitySet.create_managed_availability_set_response(@client)
+    avail_set_params =
+      {
+        name: 'myavset1',
+        location: 'west us',
+        resource_group: 'fog-test-rg',
+        is_managed: true
+      }
+
     @availability_sets.stub :validate_params, true do
       @availability_sets.stub :create_or_update, mocked_response do
-        assert_equal @service.create_availability_set('fog-test-rg', 'myavset1', 'west us', true), mocked_response
+        assert_equal @service.create_availability_set(avail_set_params), mocked_response
       end
     end
   end
 
   def test_create_managed_availability_set_failure
     response = proc { raise MsRestAzure::AzureOperationError.new(nil, nil, 'error' => { 'message' => 'mocked exception' }) }
+    avail_set_params =
+      {
+        name: 'myavset1',
+        location: 'west us',
+        resource_group: 'fog-test-rg',
+        is_managed: true
+      }
+
     @availability_sets.stub :validate_params, true do
       @availability_sets.stub :create_or_update, response do
-        assert_raises(RuntimeError) { @service.create_availability_set('fog-test-rg', 'fog-test-as', 'west us', true) }
+        assert_raises(RuntimeError) { @service.create_availability_set(avail_set_params) }
       end
     end
   end